+++ rust/lib/src/future.rs
@@ -38,7 +38,7 @@ pub(crate) trait Env {
     async fn symlink_metadata(path: &Path) -> Result<Metadata>;
     async fn open(path: &Path) -> Result<Self::File>;
     async fn ort_session_run(
-        session: &ort::Session, input: Array2<f32>,
+        session: &ort::Session, input: Array2<i32>,
     ) -> Result<ort::SessionOutputs>;
 }
 
@@ -55,7 +55,7 @@ impl Env for SyncEnv {
     }
 
     async fn ort_session_run(
-        session: &ort::Session, input: Array2<f32>,
+        session: &ort::Session, input: Array2<i32>,
     ) -> Result<ort::SessionOutputs> {
         Ok(session.run(ort::inputs!("bytes" => input)?)?)
     }
@@ -74,7 +74,7 @@ impl Env for AsyncEnv {
     }
 
     async fn ort_session_run(
-        session: &ort::Session, input: Array2<f32>,
+        session: &ort::Session, input: Array2<i32>,
     ) -> Result<ort::SessionOutputs> {
         Ok(session.run_async(ort::inputs!("bytes" => input)?)?.await?)
     }
